Add baseline utilities module and refactor
